+++ sdk/messaging/azservicebus/example_client_test.go
@@ -4,6 +4,7 @@
 package azservicebus_test
 
 import (
+	"context"
 	"fmt"
 	"time"
 
@@ -29,6 +30,25 @@ func ExampleNewClient() {
 
 	if err != nil {
 		panic(err)
+	}
+}
+
+func ExampleClient_NewListSessionsForQueuePager() {
+	// List the IDs of sessions that currently have active messages in a
+	// session-enabled queue. The IDs are returned in pages; call NextPage until
+	// More returns false.
+	pager := client.NewListSessionsForQueuePager("exampleSessionQueue", nil)
+
+	for pager.More() {
+		page, err := pager.NextPage(context.TODO())
+
+		if err != nil {
+			panic(err)
+		}
+
+		for _, sessionID := range page.Sessions {
+			fmt.Printf("Session ID: %s\n", sessionID)
+		}
 	}
 }